refactor(cli): split the jazz run output contract out of the command
run-agent.ts had grown to 649 lines holding four unrelated jobs, and workflow.ts
was importing formatters out of it — a sibling command reaching into another
command's file for the envelope, which is the tell that the envelope was never
that command's private business.

The output contract now lives in one-shot-envelope.ts: the payload types, the
three formatters, and the exit codes. It is what every bridge and script outside
this repo is written against, so a change to a published promise now reads as
one in a diff. Closed-set flag validation moves to one-shot-flags.ts. What is
left is the command itself.

Names the exit codes together as ONE_SHOT_EXIT. Parking had added a named
constant for 2 while 0 and 1 stayed bare literals at their call sites, which is
how a fourth undocumented code eventually appears.

Renames run-agent.ts to run.ts, matching the command it implements and the
convention every neighbour already follows (mcp.ts, workflow.ts, persona.ts).
